+++ CHANGELOG.md
@@ -27,6 +27,10 @@ For more information on SemVer, please visit http://semver.org/.
 
 * Loading scripts and styles
 
+#### Bugs
+
+* Fixed potential incompatibility with WordPress Multisite menu
+
 ## Version 1.5.10
 ### 2014-07-31